@@ -23,15 +23,15 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/ZcashFoundation/zebra"
 homepage = "https://zfnd.org/zebra/"
 keywords = ["zebra", "zcash"]
-rust-version = "1.85.0"
+rust-version = "1.88"
 edition = "2021"
 
 # `cargo release` settings
 
 [workspace.dependencies]
 incrementalmerkletree = { version = "0.8.2", features = ["legacy-api"] }
-orchard = "0.11"
-sapling-crypto = "0.5"
+orchard = "0.12"
+sapling-crypto = "0.6"
 zcash_address = "0.10"
 zcash_encoding = "0.3"
 zcash_history = "0.4"
@@ -59,7 +59,7 @@ bytes = "1.10"
 chrono = { version = "0.4.40", default-features = false }
 clap = "4.5"
 color-eyre = { version = "0.6.3", default-features = false }
-console-subscriber = "0.4"
+console-subscriber = "0.5"
 criterion = "0.5"
 crossbeam-channel = "0.5.14"
 derive-getters = "0.5"
@@ -136,6 +136,9 @@ schemars = "1"
 spandoc = "0.2"
 static_assertions = "1.1"
 structopt = "0.3"
+anyhow = "1.0"
+strum = "0.27"
+strum_macros = "0.27"
 syn = "2.0.100"
 tempfile = "3.20"
 thiserror = "2.0"
@@ -168,9 +171,26 @@ uint = "0.10"
 vergen-git2 = { version = "9", default-features = false }
 x25519-dalek = "2.0.1"
 zcash_note_encryption = "0.4.1"
-zcash_script = "0.4.2"
+zcash_script = "0.5"
 config = { version = "0.15.14", features = ["toml"] }
 which = "8.0.0"
+
+# TODO
+#
+# 1. Remove the whole patch table once the upstream changes land in a release.
+# 2. Also remove the corresponding URLs in the `allow-git` table in the `deny.toml` file.
+[patch.crates-io]
+zcash_primitives = { git = "https://github.com/zcash/librustzcash", branch = "bump-zcash_script" }
+zcash_address = { git = "https://github.com/zcash/librustzcash", branch = "bump-zcash_script" }
+zcash_encoding = { git = "https://github.com/zcash/librustzcash", branch = "bump-zcash_script" }
+zcash_history = { git = "https://github.com/zcash/librustzcash", branch = "bump-zcash_script" }
+zcash_keys = { git = "https://github.com/zcash/librustzcash", branch = "bump-zcash_script" }
+zcash_proofs = { git = "https://github.com/zcash/librustzcash", branch = "bump-zcash_script" }
+zcash_transparent = { git = "https://github.com/zcash/librustzcash", branch = "bump-zcash_script" }
+zcash_protocol = { git = "https://github.com/zcash/librustzcash", branch = "bump-zcash_script" }
+
+zcash_script = { git = "https://github.com/ZcashFoundation/zcash_script", branch = "master" }
+libzcash_script = { git = "https://github.com/ZcashFoundation/zcash_script", branch = "master" }
 
 [workspace.metadata.release]
 
