@@ -2,7 +2,7 @@
 
 use std::ops::{Add, Sub};
 use thiserror::Error;
-use zcash_primitives::consensus::BlockHeight;
+use zcash_protocol::consensus::BlockHeight;
 
 use crate::{serialization::SerializationError, BoxError};
 
@@ -148,6 +148,14 @@ impl TryFrom<u32> for Height {
     }
 }
 
+impl TryFrom<i64> for Height {
+    type Error = core::num::TryFromIntError;
+
+    fn try_from(height: i64) -> Result<Self, Self::Error> {
+        u32::try_from(height).map(Height)
+    }
+}
+
 impl From<Height> for u32 {
     fn from(height: Height) -> Self {
         height.0
@@ -156,6 +164,18 @@ impl From<Height> for u32 {
 
 impl From<Height> for u64 {
     fn from(height: Height) -> Self {
+        height.0.into()
+    }
+}
+
+impl From<Height> for i64 {
+    fn from(height: Height) -> Self {
+        height.0.into()
+    }
+}
+
+impl From<&Height> for i64 {
+    fn from(height: &Height) -> Self {
         height.0.into()
     }
 }
