@@ -3,7 +3,7 @@
 //! Usage: <https://docs.rs/zcash_address/0.2.0/zcash_address/trait.TryFromAddress.html#examples>
 
 use zcash_address::unified::{self, Container};
-use zcash_primitives::consensus::NetworkType;
+use zcash_protocol::consensus::NetworkType;
 
 use crate::{parameters::NetworkKind, transparent, BoxError};
 
