@@ -9,6 +9,7 @@ mod utxo;
 
 use std::{collections::HashMap, fmt, iter, ops::AddAssign};
 
+use zcash_script::{opcode::Evaluable as _, pattern::push_num};
 use zcash_transparent::{address::TransparentAddress, bundle::TxOut};
 
 use crate::{
@@ -17,11 +18,11 @@ use crate::{
     parameters::Network,
     serialization::ZcashSerialize,
     transaction,
+    transparent::serialize::GENESIS_COINBASE_SCRIPT_SIG,
 };
 
 pub use address::Address;
 pub use script::Script;
-pub use serialize::{GENESIS_COINBASE_DATA, MAX_COINBASE_DATA_LEN, MAX_COINBASE_HEIGHT_DATA_LEN};
 pub use utxo::{
     new_ordered_outputs, new_outputs, outputs_from_utxos, utxos_from_ordered_utxos,
     CoinbaseSpendRestriction, OrderedUtxo, Utxo,
@@ -52,69 +53,10 @@ use proptest_derive::Arbitrary;
 // TODO: change type to HeightDiff
 pub const MIN_TRANSPARENT_COINBASE_MATURITY: u32 = 100;
 
-/// Extra coinbase data that identifies some coinbase transactions generated by Zebra.
-/// <https://emojipedia.org/zebra/>
-//
-// # Note
-//
-// rust-analyzer will crash in some editors when moving over an actual Zebra emoji,
-// so we encode it here. This is a known issue in emacs-lsp and other lsp implementations:
-// - https://github.com/rust-lang/rust-analyzer/issues/9121
-// - https://github.com/emacs-lsp/lsp-mode/issues/2080
-// - https://github.com/rust-lang/rust-analyzer/issues/13709
-pub const EXTRA_ZEBRA_COINBASE_DATA: &str = "z\u{1F993}";
-
 /// The rate used to calculate the dust threshold, in zatoshis per 1000 bytes.
 ///
 /// History: <https://github.com/zcash/zcash/blob/v6.10.0/src/policy/policy.h#L43-L89>
 pub const ONE_THIRD_DUST_THRESHOLD_RATE: u32 = 100;
-
-/// Arbitrary data inserted by miners into a coinbase transaction.
-//
-// TODO: rename to ExtraCoinbaseData, because height is also part of the coinbase data?
-#[derive(Clone, Eq, PartialEq)]
-#[cfg_attr(
-    any(test, feature = "proptest-impl", feature = "elasticsearch"),
-    derive(Serialize)
-)]
-pub struct CoinbaseData(
-    /// Invariant: this vec, together with the coinbase height, must be less than
-    /// 100 bytes. We enforce this by only constructing CoinbaseData fields by
-    /// parsing blocks with 100-byte data fields, and checking newly created
-    /// CoinbaseData lengths in the transaction builder.
-    pub(super) Vec<u8>,
-);
-
-#[cfg(any(test, feature = "proptest-impl"))]
-impl CoinbaseData {
-    /// Create a new `CoinbaseData` containing `data`.
-    ///
-    /// Only for use in tests.
-    pub fn new(data: Vec<u8>) -> CoinbaseData {
-        CoinbaseData(data)
-    }
-}
-
-impl AsRef<[u8]> for CoinbaseData {
-    fn as_ref(&self) -> &[u8] {
-        self.0.as_ref()
-    }
-}
-
-impl std::fmt::Debug for CoinbaseData {
-    #[allow(clippy::unwrap_in_result)]
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let escaped = String::from_utf8(
-            self.0
-                .iter()
-                .cloned()
-                .flat_map(std::ascii::escape_default)
-                .collect(),
-        )
-        .expect("ascii::escape_default produces utf8");
-        f.debug_tuple("CoinbaseData").field(&escaped).finish()
-    }
-}
 
 /// OutPoint
 ///
@@ -177,8 +119,9 @@ pub enum Input {
     Coinbase {
         /// The height of this block.
         height: block::Height,
-        /// Free data inserted by miners after the block height.
-        data: CoinbaseData,
+        /// Optional, arbitrary data miners can insert into a coinbase tx.
+        /// Limited to ~ 94 bytes.
+        data: Vec<u8>,
         /// The sequence number for the output.
         sequence: u32,
     },
@@ -203,7 +146,7 @@ impl fmt::Display for Input {
                 let mut fmter = f.debug_struct("transparent::Input::Coinbase");
 
                 fmter.field("height", height);
-                fmter.field("data_len", &data.0.len());
+                fmter.field("data_len", &data.len());
 
                 fmter.finish()
             }
@@ -212,68 +155,30 @@ impl fmt::Display for Input {
 }
 
 impl Input {
-    /// Returns a new coinbase input for `height` with optional `data` and `sequence`.
-    ///
-    /// # Consensus
-    ///
-    /// The combined serialized size of `height` and `data` can be at most 100 bytes.
-    ///
-    /// > A coinbase transaction script MUST have length in {2 .. 100} bytes.
-    ///
-    /// <https://zips.z.cash/protocol/protocol.pdf#txnconsensus>
-    ///
-    /// # Panics
-    ///
-    /// If the coinbase data is greater than [`MAX_COINBASE_DATA_LEN`].
-    pub fn new_coinbase(height: block::Height, data: Vec<u8>, sequence: Option<u32>) -> Input {
-        // `zcashd` includes an extra byte after the coinbase height in the coinbase data. We do
-        // that only if the data is empty to stay compliant with the following consensus rule:
-        //
-        // > A coinbase transaction script MUST have length in {2 .. 100} bytes.
-        //
-        // ## Rationale
-        //
-        // Coinbase heights < 17 are serialized as a single byte, and if there is no coinbase data,
-        // the script of a coinbase tx with such a height would consist only of this single byte,
-        // violating the consensus rule.
-        let data = if data.is_empty() { vec![0] } else { data };
-        let data_limit = MAX_COINBASE_DATA_LEN - height.coinbase_zcash_serialized_size();
-
-        assert!(
-            data.len() <= data_limit,
-            "miner data has {} bytes, which exceeds the limit of {data_limit} bytes",
-            data.len(),
-        );
-
-        Input::Coinbase {
-            height,
-            data: CoinbaseData(data),
-            // If the caller does not specify the sequence number, use a sequence number that
-            // activates the LockTime.
-            sequence: sequence.unwrap_or(0),
-        }
-    }
-
-    /// Returns the extra coinbase data in this input, if it is an [`Input::Coinbase`].
-    pub fn extra_coinbase_data(&self) -> Option<&CoinbaseData> {
+    /// Returns the miner data in this input, if it is an [`Input::Coinbase`].
+    pub fn miner_data(&self) -> Option<&Vec<u8>> {
         match self {
-            Input::PrevOut { .. } => None,
             Input::Coinbase { data, .. } => Some(data),
+            _ => None,
         }
     }
 
-    /// Returns the full coinbase script (the encoded height along with the
-    /// extra data) if this is an [`Input::Coinbase`]. Also returns `None` if
-    /// the coinbase is for the genesis block but does not match the expected
-    /// genesis coinbase data.
+    /// Returns the full coinbase script (the encoded height along with the optional miner data) if
+    /// this is an [`Input::Coinbase`]. Also returns `None` if the coinbase is for the genesis block
+    /// but does not match the expected genesis coinbase data.
     pub fn coinbase_script(&self) -> Option<Vec<u8>> {
         match self {
             Input::PrevOut { .. } => None,
             Input::Coinbase { height, data, .. } => {
-                let mut height_and_data = Vec::new();
-                serialize::write_coinbase_height(*height, data, &mut height_and_data).ok()?;
-                height_and_data.extend(&data.0);
-                Some(height_and_data)
+                if height.is_min() {
+                    (data.as_slice() == GENESIS_COINBASE_SCRIPT_SIG)
+                        .then_some(GENESIS_COINBASE_SCRIPT_SIG.to_vec())
+                } else {
+                    let mut script = push_num(height.into()).to_bytes();
+                    script.extend_from_slice(data);
+
+                    Some(script)
+                }
             }
         }
     }
