@@ -11,6 +11,7 @@ use crate::{
 #[cfg(test)]
 use proptest::prelude::*;
 use zcash_address::{ToAddress, ZcashAddress};
+use zcash_transparent::address::TransparentAddress;
 
 /// Transparent Zcash Addresses
 ///
@@ -25,7 +26,7 @@ use zcash_address::{ToAddress, ZcashAddress};
 /// <https://zips.z.cash/protocol/protocol.pdf#transparentaddrencoding>
 // TODO Remove this type and move to `TransparentAddress` in `zcash-transparent`.
 #[derive(
-    Clone, Eq, PartialEq, Hash, serde_with::SerializeDisplay, serde_with::DeserializeFromStr,
+    Clone, Copy, Eq, PartialEq, Hash, serde_with::SerializeDisplay, serde_with::DeserializeFromStr,
 )]
 pub enum Address {
     /// P2SH (Pay to Script Hash) addresses
@@ -71,6 +72,20 @@ impl From<Address> for ZcashAddress {
                 network_kind,
                 validating_key_hash,
             } => ZcashAddress::from_tex(network_kind.into(), validating_key_hash),
+        }
+    }
+}
+
+impl TryFrom<Address> for TransparentAddress {
+    type Error = &'static str;
+
+    fn try_from(taddr: Address) -> Result<Self, Self::Error> {
+        match taddr {
+            Address::PayToScriptHash { script_hash, .. } => Ok(Self::ScriptHash(script_hash)),
+            Address::PayToPublicKeyHash { pub_key_hash, .. } => {
+                Ok(Self::PublicKeyHash(pub_key_hash))
+            }
+            Address::Tex { .. } => Err("TransparentAddress can't be a Tex address"),
         }
     }
 }
@@ -142,12 +157,12 @@ impl std::str::FromStr for Address {
         hash_bytes.copy_from_slice(&payload);
 
         match hrp.as_str() {
-            zcash_primitives::constants::mainnet::HRP_TEX_ADDRESS => Ok(Address::Tex {
+            zcash_protocol::constants::mainnet::HRP_TEX_ADDRESS => Ok(Address::Tex {
                 network_kind: NetworkKind::Mainnet,
                 validating_key_hash: hash_bytes,
             }),
 
-            zcash_primitives::constants::testnet::HRP_TEX_ADDRESS => Ok(Address::Tex {
+            zcash_protocol::constants::testnet::HRP_TEX_ADDRESS => Ok(Address::Tex {
                 network_kind: NetworkKind::Testnet,
                 validating_key_hash: hash_bytes,
             }),
@@ -196,25 +211,25 @@ impl ZcashDeserialize for Address {
         reader.read_exact(&mut hash_bytes)?;
 
         match version_bytes {
-            zcash_primitives::constants::mainnet::B58_SCRIPT_ADDRESS_PREFIX => {
+            zcash_protocol::constants::mainnet::B58_SCRIPT_ADDRESS_PREFIX => {
                 Ok(Address::PayToScriptHash {
                     network_kind: NetworkKind::Mainnet,
                     script_hash: hash_bytes,
                 })
             }
-            zcash_primitives::constants::testnet::B58_SCRIPT_ADDRESS_PREFIX => {
+            zcash_protocol::constants::testnet::B58_SCRIPT_ADDRESS_PREFIX => {
                 Ok(Address::PayToScriptHash {
                     network_kind: NetworkKind::Testnet,
                     script_hash: hash_bytes,
                 })
             }
-            zcash_primitives::constants::mainnet::B58_PUBKEY_ADDRESS_PREFIX => {
+            zcash_protocol::constants::mainnet::B58_PUBKEY_ADDRESS_PREFIX => {
                 Ok(Address::PayToPublicKeyHash {
                     network_kind: NetworkKind::Mainnet,
                     pub_key_hash: hash_bytes,
                 })
             }
-            zcash_primitives::constants::testnet::B58_PUBKEY_ADDRESS_PREFIX => {
+            zcash_protocol::constants::testnet::B58_PUBKEY_ADDRESS_PREFIX => {
                 Ok(Address::PayToPublicKeyHash {
                     network_kind: NetworkKind::Testnet,
                     pub_key_hash: hash_bytes,
