@@ -33,7 +33,7 @@ use zebra_chain::{
         },
         zip317, Hash, HashType, JoinSplitData, LockTime, Transaction,
     },
-    transparent::{self, CoinbaseData, CoinbaseSpendRestriction},
+    transparent::{self, CoinbaseSpendRestriction},
 };
 
 use zebra_node_services::mempool;
@@ -3116,7 +3116,7 @@ fn mock_coinbase_transparent_output(
 
     let input = transparent::Input::Coinbase {
         height: coinbase_height,
-        data: CoinbaseData::new(Vec::new()),
+        data: vec![],
         sequence: u32::MAX,
     };
 
