@@ -1007,7 +1007,8 @@ pub async fn test_mining_rpcs<State, ReadState>(
             NetworkType::from(NetworkKind::from(network)),
             [0x7e; 20],
         )),
-        extra_coinbase_data: None,
+        miner_data: None,
+        miner_memo: None,
         // TODO: Use default field values when optional features are enabled in tests #8183
         internal_miner: true,
     };
