@@ -1,45 +1,97 @@
 //! Tests for types and functions for the `getblocktemplate` RPC.
 
+use anyhow::anyhow;
+use std::iter;
+
+use strum::IntoEnumIterator;
 use zcash_keys::address::Address;
-use zcash_transparent::address::TransparentAddress;
+
+use zebra_chain::parameters::testnet::ConfiguredFundingStreamRecipient;
 
 use zebra_chain::{
-    amount::Amount,
     block::Height,
-    parameters::testnet::{self, ConfiguredActivationHeights, ConfiguredFundingStreams},
-    serialization::{ZcashDeserializeInto, ZcashSerialize},
+    parameters::{
+        subsidy::FundingStreamReceiver::{Deferred, Ecc, MajorGrants, ZcashFoundation},
+        testnet::{self, ConfiguredActivationHeights, ConfiguredFundingStreams},
+        Network, NetworkUpgrade,
+    },
+    serialization::ZcashDeserializeInto,
     transaction::Transaction,
 };
 
-use super::standard_coinbase_outputs;
+use crate::client::TransactionTemplate;
+use crate::config::mining::{MinerAddressType, MINER_ADDRESS};
 
-/// Tests that a minimal coinbase transaction can be generated.
+use super::MinerParams;
+
+/// Tests that coinbase transactions can be generated.
+///
+/// This test needs to be run with the `--release` flag so that it runs for ~ 30 seconds instead of
+/// ~ 90.
 #[test]
-fn minimal_coinbase() -> Result<(), Box<dyn std::error::Error>> {
+#[ignore]
+fn coinbase() -> anyhow::Result<()> {
     let regtest = testnet::Parameters::build()
         .with_slow_start_interval(Height::MIN)
         .with_activation_heights(ConfiguredActivationHeights {
-            nu6: Some(1),
+            overwinter: Some(1),
+            sapling: Some(2),
+            blossom: Some(3),
+            heartwood: Some(4),
+            canopy: Some(5),
+            nu5: Some(6),
+            nu6: Some(7),
+            nu6_1: Some(8),
+            nu7: Some(9),
             ..Default::default()
         })?
-        .with_funding_streams(vec![ConfiguredFundingStreams {
-            height_range: Some(Height(1)..Height(10)),
-            recipients: None,
-        }])
+        .with_funding_streams(vec![
+            ConfiguredFundingStreams {
+                height_range: Some(Height(1)..Height(100)),
+                recipients: Some(vec![
+                    ConfiguredFundingStreamRecipient::new_for(Ecc),
+                    ConfiguredFundingStreamRecipient::new_for(ZcashFoundation),
+                    ConfiguredFundingStreamRecipient::new_for(MajorGrants),
+                ]),
+            },
+            ConfiguredFundingStreams {
+                height_range: Some(Height(1)..Height(100)),
+                recipients: Some(vec![
+                    ConfiguredFundingStreamRecipient::new_for(MajorGrants),
+                    ConfiguredFundingStreamRecipient {
+                        receiver: Deferred,
+                        numerator: 12,
+                        addresses: None,
+                    },
+                ]),
+            },
+        ])
         .to_network()?;
 
-    let outputs = standard_coinbase_outputs(
-        &regtest,
-        Height(1),
-        &Address::from(TransparentAddress::PublicKeyHash([0x42; 20])),
-        Amount::zero(),
-    );
-
-    // It should be possible to generate a coinbase tx from these params.
-    Transaction::new_v5_coinbase(&regtest, Height(1), outputs, vec![])
-        .zcash_serialize_to_vec()?
-        // Deserialization contains checks for elementary consensus rules, which must pass.
-        .zcash_deserialize_into::<Transaction>()?;
+    for net in Network::iter().chain(iter::once(regtest)) {
+        for nu in NetworkUpgrade::iter().filter(|nu| nu >= &NetworkUpgrade::Sapling) {
+            if let Some(height) = nu.activation_height(&net) {
+                for addr_type in MinerAddressType::iter() {
+                    TransactionTemplate::new_coinbase(
+                        &net,
+                        height,
+                        &MinerParams::from(
+                            Address::decode(&net, MINER_ADDRESS[&net.kind()][&addr_type])
+                                .ok_or(anyhow!("hard-coded addr must be valid"))?,
+                        ),
+                        &[],
+                        #[cfg(all(zcash_unstable = "nu7", feature = "tx_v6"))]
+                        None,
+                    )?
+                    .data()
+                    .as_ref()
+                    // Deserialization contains checks for elementary consensus rules, which must
+                    // pass.
+                    .zcash_deserialize_into::<Transaction>()?;
+                }
+            }
+        }
+    }
 
     Ok(())
 }
