@@ -13,13 +13,10 @@ use rand::{
     prelude::thread_rng,
 };
 
-use zcash_keys::address::Address;
-
 use zebra_chain::{
-    amount::NegativeOrZero,
     block::{Height, MAX_BLOCK_BYTES},
     parameters::Network,
-    transaction::{self, zip317::BLOCK_UNPAID_ACTION_LIMIT, Transaction, VerifiedUnminedTx},
+    transaction::{self, zip317::BLOCK_UNPAID_ACTION_LIMIT, VerifiedUnminedTx},
 };
 use zebra_consensus::MAX_BLOCK_SIGOPS;
 use zebra_node_services::mempool::TransactionDependencies;
@@ -29,16 +26,13 @@ use crate::methods::types::transaction::TransactionTemplate;
 #[cfg(all(zcash_unstable = "nu7", feature = "tx_v6"))]
 use crate::methods::{Amount, NonNegative};
 
-#[cfg(all(zcash_unstable = "nu7", feature = "tx_v6"))]
-use zebra_chain::parameters::NetworkUpgrade;
-
 #[cfg(test)]
 mod tests;
 
 #[cfg(test)]
 use crate::methods::types::get_block_template::InBlockTxDependenciesDepth;
 
-use super::standard_coinbase_outputs;
+use super::MinerParams;
 
 /// Used in the return type of [`select_mempool_transactions()`] for test compilations.
 #[cfg(test)]
@@ -60,26 +54,26 @@ type SelectedMempoolTx = VerifiedUnminedTx;
 /// [ZIP-317]: https://zips.z.cash/zip-0317#block-production
 #[allow(clippy::too_many_arguments)]
 pub fn select_mempool_transactions(
-    network: &Network,
-    next_block_height: Height,
-    miner_address: &Address,
+    net: &Network,
+    height: Height,
+    miner_params: &MinerParams,
     mempool_txs: Vec<VerifiedUnminedTx>,
     mempool_tx_deps: TransactionDependencies,
-    extra_coinbase_data: Vec<u8>,
     #[cfg(all(zcash_unstable = "nu7", feature = "tx_v6"))] zip233_amount: Option<
         Amount<NonNegative>,
     >,
 ) -> Vec<SelectedMempoolTx> {
     // Use a fake coinbase transaction to break the dependency between transaction
     // selection, the miner fee, and the fee payment in the coinbase transaction.
-    let fake_coinbase_tx = fake_coinbase_transaction(
-        network,
-        next_block_height,
-        miner_address,
-        extra_coinbase_data,
+    let fake_coinbase_tx = TransactionTemplate::new_coinbase(
+        net,
+        height,
+        miner_params,
+        &[],
         #[cfg(all(zcash_unstable = "nu7", feature = "tx_v6"))]
         zip233_amount,
-    );
+    )
+    .expect("valid coinbase transaction template");
 
     let tx_dependencies = mempool_tx_deps.dependencies();
     let (independent_mempool_txs, mut dependent_mempool_txs): (HashMap<_, _>, HashMap<_, _>) =
@@ -140,51 +134,6 @@ pub fn select_mempool_transactions(
     }
 
     selected_txs
-}
-
-/// Returns a fake coinbase transaction that can be used during transaction selection.
-///
-/// This avoids a data dependency loop involving the selected transactions, the miner fee,
-/// and the coinbase transaction.
-///
-/// This transaction's serialized size and sigops must be at least as large as the real coinbase
-/// transaction with the correct height and fee.
-pub fn fake_coinbase_transaction(
-    net: &Network,
-    height: Height,
-    miner_address: &Address,
-    extra_coinbase_data: Vec<u8>,
-    #[cfg(all(zcash_unstable = "nu7", feature = "tx_v6"))] zip233_amount: Option<
-        Amount<NonNegative>,
-    >,
-) -> TransactionTemplate<NegativeOrZero> {
-    // Block heights are encoded as variable-length (script) and `u32` (lock time, expiry height).
-    // They can also change the `u32` consensus branch id.
-    // We use the template height here, which has the correct byte length.
-    // https://zips.z.cash/protocol/protocol.pdf#txnconsensus
-    // https://github.com/zcash/zips/blob/main/zip-0203.rst#changes-for-nu5
-    //
-    // Transparent amounts are encoded as `i64`,
-    // so one zat has the same size as the real amount:
-    // https://developer.bitcoin.org/reference/transactions.html#txout-a-transaction-output
-    let miner_fee = 1.try_into().expect("amount is valid and non-negative");
-    let outputs = standard_coinbase_outputs(net, height, miner_address, miner_fee);
-
-    #[cfg(not(all(zcash_unstable = "nu7", feature = "tx_v6")))]
-    let coinbase = Transaction::new_v5_coinbase(net, height, outputs, extra_coinbase_data).into();
-
-    #[cfg(all(zcash_unstable = "nu7", feature = "tx_v6"))]
-    let coinbase = {
-        let network_upgrade = NetworkUpgrade::current(net, height);
-        if network_upgrade < NetworkUpgrade::Nu7 {
-            Transaction::new_v5_coinbase(net, height, outputs, extra_coinbase_data).into()
-        } else {
-            Transaction::new_v6_coinbase(net, height, outputs, extra_coinbase_data, zip233_amount)
-                .into()
-        }
-    };
-
-    TransactionTemplate::from_coinbase(&coinbase, miner_fee)
 }
 
 /// Returns a fee-weighted index and the total weight of `transactions`.
