@@ -3,20 +3,19 @@
 #![allow(clippy::unwrap_in_result)]
 
 use zcash_keys::address::Address;
-
 use zcash_transparent::address::TransparentAddress;
+
 use zebra_chain::{block::Height, parameters::Network, transaction, transparent::OutPoint};
 use zebra_node_services::mempool::TransactionDependencies;
+
+use crate::methods::types::get_block_template::MinerParams;
 
 use super::select_mempool_transactions;
 
 #[test]
 fn excludes_tx_with_unselected_dependencies() {
     let network = Network::Mainnet;
-    let next_block_height = Height(1_000_000);
-    let extra_coinbase_data = Vec::new();
     let mut mempool_tx_deps = TransactionDependencies::default();
-    let miner_address = Address::from(TransparentAddress::PublicKeyHash([0x7e; 20]));
 
     let unmined_tx = network
         .unmined_transactions_in_blocks(..)
@@ -31,11 +30,10 @@ fn excludes_tx_with_unselected_dependencies() {
     assert_eq!(
         select_mempool_transactions(
             &network,
-            next_block_height,
-            &miner_address,
+            Height(1_000_000),
+            &MinerParams::from(Address::from(TransparentAddress::PublicKeyHash([0x7e; 20]))),
             vec![unmined_tx],
             mempool_tx_deps,
-            extra_coinbase_data,
             #[cfg(all(zcash_unstable = "nu7", feature = "tx_v6"))]
             None,
         ),
@@ -47,9 +45,7 @@ fn excludes_tx_with_unselected_dependencies() {
 #[test]
 fn includes_tx_with_selected_dependencies() {
     let network = Network::Mainnet;
-    let next_block_height = Height(1_000_000);
     let unmined_txs: Vec<_> = network.unmined_transactions_in_blocks(..).take(3).collect();
-    let miner_address = Address::from(TransparentAddress::PublicKeyHash([0x7e; 20]));
 
     let dependent_tx1 = unmined_txs.first().expect("should have 3 txns");
     let dependent_tx2 = unmined_txs.get(1).expect("should have 3 txns");
@@ -73,15 +69,12 @@ fn includes_tx_with_selected_dependencies() {
         ],
     );
 
-    let extra_coinbase_data = Vec::new();
-
     let selected_txs = select_mempool_transactions(
         &network,
-        next_block_height,
-        &miner_address,
+        Height(1_000_000),
+        &MinerParams::from(Address::from(TransparentAddress::PublicKeyHash([0x7e; 20]))),
         unmined_txs.clone(),
         mempool_tx_deps.clone(),
-        extra_coinbase_data,
         #[cfg(all(zcash_unstable = "nu7", feature = "tx_v6"))]
         None,
     );
