@@ -910,7 +910,7 @@ fn test_get_address_utxos_chain_info_false() -> Result<(), Box<dyn std::error::E
         .iter()
         .map(|utxo| {
             // Address extractability was checked manually
-            let address = utxo.address().clone();
+            let address = utxo.address();
             // Hash extractability was checked in other test
             let txid = utxo.txid();
             let output_index = utxo.output_index().index();
@@ -921,7 +921,7 @@ fn test_get_address_utxos_chain_info_false() -> Result<(), Box<dyn std::error::E
             let height = utxo.height();
 
             Utxo::new(
-                address,
+                *address,
                 txid,
                 OutputIndex::from_index(output_index),
                 script,
@@ -968,7 +968,7 @@ fn test_get_address_utxos_chain_info_true() -> Result<(), Box<dyn std::error::Er
             .iter()
             .map(|utxo| {
                 // Address extractability was checked manually
-                let address = utxo.address().clone();
+                let address = utxo.address();
                 // Hash extractability was checked in other test
                 let txid = utxo.txid();
                 let output_index = utxo.output_index().index();
@@ -979,7 +979,7 @@ fn test_get_address_utxos_chain_info_true() -> Result<(), Box<dyn std::error::Er
                 let height = utxo.height();
 
                 Utxo::new(
-                    address,
+                    *address,
                     txid,
                     OutputIndex::from_index(output_index),
                     script,
@@ -1396,9 +1396,9 @@ fn test_get_block_subsidy() -> Result<(), Box<dyn std::error::Error>> {
             let specification = stream.specification().clone();
             let value = stream.value();
             let value_zat = stream.value_zat();
-            let address = stream.address().clone();
+            let address = stream.address();
 
-            FundingStream::new(recipient, specification, value, value_zat, address)
+            FundingStream::new(recipient, specification, value, value_zat, *address)
         })
         .collect::<Vec<_>>();
     let lockbox_streams = obj.lockbox_streams().clone();
