@@ -9,10 +9,7 @@ use tokio::sync::{
 };
 
 use tracing::Span;
-use zebra_chain::{
-    block::{self, Height},
-    transparent::EXTRA_ZEBRA_COINBASE_DATA,
-};
+use zebra_chain::block::{self, Height};
 
 use crate::{
     constants::MAX_BLOCK_REORG_HEIGHT,
@@ -83,7 +80,6 @@ pub(crate) fn validate_and_commit_non_finalized(
         non_finalized_state,
         chain_tip_sender,
         non_finalized_state_sender,
-        last_zebra_mined_log_height
     ),
     fields(chains = non_finalized_state.chain_count())
 )]
@@ -91,7 +87,6 @@ fn update_latest_chain_channels(
     non_finalized_state: &NonFinalizedState,
     chain_tip_sender: &mut ChainTipSender,
     non_finalized_state_sender: &watch::Sender<NonFinalizedState>,
-    last_zebra_mined_log_height: &mut Option<Height>,
 ) -> block::Height {
     let best_chain = non_finalized_state.best_chain().expect("unexpected empty non-finalized state: must commit at least one block before updating channels");
 
@@ -100,8 +95,6 @@ fn update_latest_chain_channels(
         .expect("unexpected empty chain: must commit at least one block before updating channels")
         .clone();
     let tip_block = ChainTipBlock::from(tip_block);
-
-    log_if_mined_by_zebra(&tip_block, last_zebra_mined_log_height);
 
     let tip_block_height = tip_block.height;
 
@@ -246,7 +239,6 @@ impl WriteBlockWorkerTask {
             non_finalized_state_sender,
         } = &mut self;
 
-        let mut last_zebra_mined_log_height = None;
         let mut prev_finalized_note_commitment_trees = None;
 
         // Write all the finalized blocks sent by the state,
@@ -292,9 +284,6 @@ impl WriteBlockWorkerTask {
                 Ok((finalized, note_commitment_trees)) => {
                     let tip_block = ChainTipBlock::from(finalized);
                     prev_finalized_note_commitment_trees = Some(note_commitment_trees);
-
-                    log_if_mined_by_zebra(&tip_block, &mut last_zebra_mined_log_height);
-
                     chain_tip_sender.set_finalized_tip(tip_block);
                 }
                 Err(error) => {
@@ -354,7 +343,6 @@ impl WriteBlockWorkerTask {
                     non_finalized_state,
                     chain_tip_sender,
                     non_finalized_state_sender,
-                    &mut last_zebra_mined_log_height,
                 );
                 continue;
             };
@@ -410,7 +398,6 @@ impl WriteBlockWorkerTask {
                 non_finalized_state,
                 chain_tip_sender,
                 non_finalized_state_sender,
-                &mut last_zebra_mined_log_height,
             );
 
             // Update the caller with the result.
@@ -451,74 +438,5 @@ impl WriteBlockWorkerTask {
         // done writing to the finalized state, so we can force it to shut down.
         finalized_state.db.shutdown(true);
         std::mem::drop(self.finalized_state);
-    }
-}
-
-/// Log a message if this block was mined by Zebra.
-///
-/// Does not detect early Zebra blocks, and blocks with custom coinbase transactions.
-/// Rate-limited to every 1000 blocks using `last_zebra_mined_log_height`.
-fn log_if_mined_by_zebra(
-    tip_block: &ChainTipBlock,
-    last_zebra_mined_log_height: &mut Option<Height>,
-) {
-    // This logs at most every 2-3 checkpoints, which seems fine.
-    const LOG_RATE_LIMIT: u32 = 1000;
-
-    let height = tip_block.height.0;
-
-    if let Some(last_height) = last_zebra_mined_log_height {
-        if height < last_height.0 + LOG_RATE_LIMIT {
-            // If we logged in the last 1000 blocks, don't log anything now.
-            return;
-        }
-    };
-
-    // This code is rate-limited, so we can do expensive transformations here.
-    let coinbase_data = tip_block.transactions[0].inputs()[0]
-        .extra_coinbase_data()
-        .expect("valid blocks must start with a coinbase input")
-        .clone();
-
-    if coinbase_data
-        .as_ref()
-        .starts_with(EXTRA_ZEBRA_COINBASE_DATA.as_bytes())
-    {
-        let text = String::from_utf8_lossy(coinbase_data.as_ref());
-
-        *last_zebra_mined_log_height = Some(Height(height));
-
-        // No need for hex-encoded data if it's exactly what we expected.
-        if coinbase_data.as_ref() == EXTRA_ZEBRA_COINBASE_DATA.as_bytes() {
-            info!(
-                %text,
-                %height,
-                hash = %tip_block.hash,
-                "looks like this block was mined by Zebra!"
-            );
-        } else {
-            // # Security
-            //
-            // Use the extra data as an allow-list, replacing unknown characters.
-            // This makes sure control characters and harmful messages don't get logged
-            // to the terminal.
-            let text = text.replace(
-                |c: char| {
-                    !EXTRA_ZEBRA_COINBASE_DATA
-                        .to_ascii_lowercase()
-                        .contains(c.to_ascii_lowercase())
-                },
-                "?",
-            );
-            let data = hex::encode(coinbase_data.as_ref());
-
-            info!(
-                %text,
-                %data,
-                %height,
-                hash = %tip_block.hash,
-                "looks like this block was mined by Zebra!"
-            );
-        }
     }
 }
